Revert "Refactor bot.py into smaller files"
